V1.1 staging (#300)
* feat: forceIterateOverTicks (#1)

* initial commit

* save

* save

* fix: use $clearingPrice for submitBid

* forge fmt

* save

* fix invariant

* rm output.txt

* chore: forge fmt

* fix: add onlyActiveAuction check to forceIterateOverTicks

* test: add btt for forceIterateOverTicks

* test: update btt

* Add dos test

* test: lower fuzz runs

* feat: move clearing price initialization to constructor

* test: fix invariant test to roll to start block

* feedback

* feedback: save new clearing price on stack

* gas: only update clearingPrice if changed and emit event

* chore: forge fmt

* chore: update interface

* chore: add comments from last SB audit and pre-commit (#3)

* chore: standardize parameter prefix _ (#7)

* chore: fix missing interface getter

* feat: add BlockNumberish to support arb block numbers (#2)

* feat: add blockNumberish to CCA.sol

* chore: dedupe calls

* add blocknumberish library

* remove BlockNumberish contract from local src

* remove IArbSys interface

* chore: gas

* fix: missing block numberish in events

* chore: update gas

* periphery: Add gated 1155 hooks (#5)

* initial commit of gated 1155 hooks

* chore: add tests

* chore: update pre-commit docs (#8)

* lower optimizer runs for factory (#9)

* feat: implement ILBPInitializer interface (#4)

* feat: implement ILBPInitializer interface

* fix: use global clearing price in lbpInitializationParams

* Fully implement ILBPInitializer interface

* chore: bound block numbers to uint64 max (#11)

* fix: batch claim error params order (#13)

* fix: use blocknumberish for gated ERC1155 validation hook (#12)

* audit(OZ): Notes (#17)

* N02: remove constructor blocknumberish call

* N05

* chore: update readme with guides, deployments, etc. (#14)

* chore: update readme and add docs

* chore: update deployment guide

* chore: update

* chore: update changelog

* fix: add tests for lbpInitializationParams and fix revert reason (#16)

* fix: add tests for lbpInitializationParams and fix revert reason

* chore: add supportsInterface test

* rm submodule

* chore: update ILBPInitializer interface id (#18)

* audit(sb-05): prevent re-entrancy from validation hook (#10)

* make external call before reading from state

* feat: add nonReentrant on submitBid and forceIterateOverTicks, also call internal checkpoint in submitBid

* chore: add btt tests for validation hook reentrancy

* chore: forge fmt

* chore: lower pr fuzz runs & pre-commit (#19)

* chore: forge fmt

* chore: lower pr fuzz runs

* chore: pre-commit

* chore: fix test calling cheatcode in fuzz campaign

* chore: update deploy script for CCA factory (#20)

* chore: update deploy script for CCA factory

* chore: update readme

* chore: lower pr and ci fuzz runs

* chore: fix readme conflict

+++ test/btt/auction/SweepUnsoldTokens.t.sol
@@ -44,7 +44,7 @@ contract SweepUnsoldTokensTest is BttBase {
         MockContinuousClearingAuction auction =
             new MockContinuousClearingAuction(mParams.token, mParams.totalSupply, mParams.parameters);
 
-        uint256 blockNumber = bound(_blockNumber, mParams.parameters.endBlock, type(uint256).max);
+        uint256 blockNumber = bound(_blockNumber, mParams.parameters.endBlock, type(uint64).max);
 
         ERC20Mock(mParams.token).mint(address(auction), mParams.totalSupply);
         auction.onTokensReceived();
