+++ test/btt/auction/onlyAfterClaimBlock.t.sol
@@ -34,7 +34,7 @@ contract OnlyAfterClaimBlockTest is BttBase {
         MockContinuousClearingAuction auction =
             new MockContinuousClearingAuction(mParams.token, mParams.totalSupply, mParams.parameters);
 
-        uint256 blockNumber = bound(_blockNumber, mParams.parameters.claimBlock, type(uint256).max);
+        uint256 blockNumber = bound(_blockNumber, mParams.parameters.claimBlock, type(uint64).max);
 
         vm.roll(blockNumber);
         auction.modifier_onlyAfterClaimBlock();
